+++ src/main/java/com/herokuapp/runningappbackend/model/UserChallenge.java
@@ -32,7 +32,7 @@ public class UserChallenge {
     private Challenge challenge;
 
     @Column(name = "current_amount")
-    private int currentAmount;
+    private Double currentAmount;
 
     @Column(name = "join_date")
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss",
@@ -46,7 +46,7 @@ public class UserChallenge {
 
     private Boolean isCompleted;
 
-    public UserChallenge(User user, Challenge challenge, int currentAmount, LocalDateTime joinDate) {
+    public UserChallenge(User user, Challenge challenge, Double currentAmount, LocalDateTime joinDate) {
         this.id = new UserChallengeId(user.getId(), challenge.getId());
         this.user = user;
         this.challenge = challenge;
